Refine signing interfaces and rename signer packages
