+++ pkgs/pyproject.toml
@@ -144,10 +144,9 @@ members = [
     "standards/swarmauri_tokens_dpopboundjwt",
     "standards/swarmauri_signing_pgp",
     "standards/swarmauri_signing_rsa",
-    "standards/swarmauri_signer_cms",
-    "standards/swarmauri_signer_jws",
-    "standards/swarmauri_signer_openpgp",
-    "standards/swarmauri_signer_pades",
+    "standards/swarmauri_signing_cms",
+    "standards/swarmauri_signing_openpgp",
+    "standards/swarmauri_signing_pades",
     "swarmauri_standards",
     "standards/swarmauri_signing_ssh",
     "standards/swarmauri_certs_x509verify",
