--- pkgs/standards/swarmauri_signer_jws/pyproject.toml
@@ -1,20 +1,0 @@
-[project]
-name = "swarmauri_signer_jws"
-version = "0.1.0.dev0"
-description = "JWS signer stub registered on the Swarmauri SigningBase registry"
-license = "Apache-2.0"
-readme = "README.md"
-requires-python = ">=3.10,<3.13"
-authors = [{ name = "Jacob Stewart", email = "jacob@swarmauri.com" }]
-dependencies = [
-    "swarmauri_base",
-    "swarmauri_core",
-]
-
-[tool.uv.sources]
-swarmauri_base = { workspace = true }
-swarmauri_core = { workspace = true }
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
