--- pkgs/standards/swarmauri_signer_jws/swarmauri_signer_jws/__init__.py
@@ -1,5 +1,0 @@
-"""JWS signing plugin package."""
-
-from .jws_signer import JWSSigner
-
-__all__ = ["JWSSigner"]
