+++ pkgs/standards/swarmauri_signing_ecdsa/swarmauri_signing_ecdsa/EcdsaEnvelopeSigner.py
@@ -186,7 +186,14 @@ class EcdsaEnvelopeSigner(SigningBase):
     def supports(self) -> Mapping[str, Iterable[str]]:
         algs = tuple(_EC_ALGS.keys()) + tuple(_ALIAS.keys())
         canons = ("json", "cbor") if _CBOR_OK else ("json",)
-        return {"algs": algs, "canons": canons, "features": ("multi", "detached_only")}
+        return {
+            "algs": algs,
+            "canons": canons,
+            "signs": ("bytes", "envelope"),
+            "verifies": ("bytes", "envelope"),
+            "envelopes": ("mapping",),
+            "features": ("multi", "detached_only"),
+        }
 
     async def sign_bytes(
         self,
