+++ pkgs/standards/swarmauri_signing_jws/swarmauri_signing_jws/__init__.py
@@ -1,3 +1,4 @@
 from .JwsSignerVerifier import JwsSignerVerifier, JwsResult
+from .jws_signer import JWSSigner
 
-__all__ = ["JwsSignerVerifier", "JwsResult"]
+__all__ = ["JwsSignerVerifier", "JwsResult", "JWSSigner"]
