+++ pkgs/standards/swarmauri_signing_openpgp/LICENSE
