+++ pkgs/standards/swarmauri_signing_pades/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "swarmauri_signer_openpgp"
+name = "swarmauri_signing_pades"
 version = "0.1.0.dev0"
-description = "OpenPGP signer stub registered on the Swarmauri SigningBase registry"
+description = "PAdES signer stub registered on the Swarmauri SigningBase registry"
 license = "Apache-2.0"
 readme = "README.md"
 requires-python = ">=3.10,<3.13"
