+++ pkgs/standards/swarmauri_signing_pgp/swarmauri_signing_pgp/pgp_signer.py
@@ -132,6 +132,9 @@ class PgpEnvelopeSigner(SigningBase):
         return {
             "algs": ("OpenPGP",),
             "canons": canons,
+            "signs": ("bytes", "envelope"),
+            "verifies": ("bytes", "envelope"),
+            "envelopes": ("mapping",),
             "features": ("multi", "detached_only"),
         }
