+++ pkgs/standards/swarmauri_signing_rsa/swarmauri_signing_rsa/RSAEnvelopeSigner.py
@@ -244,7 +244,14 @@ class RSAEnvelopeSigner(SigningBase):
     def supports(self) -> Mapping[str, Iterable[str]]:
         algs = ("RSA-PSS-SHA256", "RS256")
         canons = ("json", "cbor") if _CBOR_OK else ("json",)
-        return {"algs": algs, "canons": canons, "features": ("multi", "detached_only")}
+        return {
+            "algs": algs,
+            "canons": canons,
+            "signs": ("bytes", "envelope"),
+            "verifies": ("bytes", "envelope"),
+            "envelopes": ("mapping",),
+            "features": ("multi", "detached_only"),
+        }
 
     async def sign_bytes(
         self,
