+++ pkgs/swarmauri_standard/swarmauri_standard/signing/__init__.py
@@ -1,0 +1,7 @@
+"""Signing facade entrypoints for swarmauri_standard."""
+
+from __future__ import annotations
+
+from .Signer import Signer
+
+__all__ = ["Signer"]
